refactor(news): rename CollectedNewsResponse to CollectedNewsApiResponse (#98)
