+++ src/main/java/com/solv/wefin/domain/news/config/EmbeddingConfig.java
@@ -1,0 +1,18 @@
+package com.solv.wefin.domain.news.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.http.client.SimpleClientHttpRequestFactory;
+import org.springframework.web.client.RestTemplate;
+
+@Configuration
+public class EmbeddingConfig {
+
+    @Bean
+    public RestTemplate embeddingRestTemplate() {
+        SimpleClientHttpRequestFactory factory = new SimpleClientHttpRequestFactory();
+        factory.setConnectTimeout(10000);
+        factory.setReadTimeout(30000);
+        return new RestTemplate(factory);
+    }
+}
