--- src/main/java/com/solv/wefin/domain/news/dto/CollectedNewsDto.java
@@ -1,20 +1,0 @@
-package com.solv.wefin.domain.news.dto;
-
-import lombok.Builder;
-import lombok.Getter;
-
-import java.time.LocalDateTime;
-
-@Getter
-@Builder
-public class CollectedNewsDto {
-
-    private String externalArticleId;
-    private String originalUrl;
-    private String originalTitle;
-    private String originalContent;
-    private String originalThumbnailUrl;
-    private LocalDateTime originalPublishedAt;
-    private String publisherName;
-    private String rawPayload;
-}
