+++ src/main/java/com/solv/wefin/domain/news/ingestion/repository/NewsCollectBatchRepository.java
@@ -1,0 +1,7 @@
+package com.solv.wefin.domain.news.ingestion.repository;
+
+import com.solv.wefin.domain.news.ingestion.entity.NewsCollectBatch;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface NewsCollectBatchRepository extends JpaRepository<NewsCollectBatch, Long> {
+}
