--- src/main/java/com/solv/wefin/domain/news/repository/RawNewsArticleRepository.java
@@ -1,13 +1,0 @@
-package com.solv.wefin.domain.news.repository;
-
-import com.solv.wefin.domain.news.entity.RawNewsArticle;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-
-public interface RawNewsArticleRepository extends JpaRepository<RawNewsArticle, Long> {
-
-    boolean existsByOriginalUrl(String originalUrl);
-
-    List<RawNewsArticle> findByProcessingStatus(RawNewsArticle.ProcessingStatus processingStatus);
-}
