+++ src/test/java/com/solv/wefin/domain/news/crawl/ArticleCrawlPersistenceServiceIntegrationTest.java
@@ -1,9 +1,10 @@
 package com.solv.wefin.domain.news.crawl;
 
 import com.solv.wefin.common.IntegrationTestBase;
-import com.solv.wefin.domain.news.entity.NewsArticle;
-import com.solv.wefin.domain.news.entity.NewsArticle.CrawlStatus;
-import com.solv.wefin.domain.news.repository.NewsArticleRepository;
+import com.solv.wefin.domain.news.ingestion.service.ArticleCrawlPersistenceService;
+import com.solv.wefin.domain.news.article.entity.NewsArticle;
+import com.solv.wefin.domain.news.article.entity.NewsArticle.CrawlStatus;
+import com.solv.wefin.domain.news.article.repository.NewsArticleRepository;
 import jakarta.persistence.EntityManager;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Test;
