+++ src/test/java/com/solv/wefin/domain/news/embedding/service/EmbeddingPersistenceServiceIntegrationTest.java
@@ -3,9 +3,9 @@ package com.solv.wefin.domain.news.embedding.service;
 import com.solv.wefin.common.IntegrationTestBase;
 import com.solv.wefin.domain.news.embedding.entity.ArticleEmbedding;
 import com.solv.wefin.domain.news.embedding.repository.ArticleEmbeddingRepository;
-import com.solv.wefin.domain.news.entity.NewsArticle;
-import com.solv.wefin.domain.news.entity.NewsArticle.EmbeddingStatus;
-import com.solv.wefin.domain.news.repository.NewsArticleRepository;
+import com.solv.wefin.domain.news.article.entity.NewsArticle;
+import com.solv.wefin.domain.news.article.entity.NewsArticle.EmbeddingStatus;
+import com.solv.wefin.domain.news.article.repository.NewsArticleRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
